+++ vortex_device/src/models/CylinderVox.cpp
@@ -9,10 +9,10 @@ CylinderVox::CylinderVox(int delta_theta){
     this->bin_size = 360/delta_theta;
 }
 
-void CylinderVox::add_point(int theta, int x_normalized, int y_normalized) {
+void CylinderVox::add_point(double theta, int x_normalized, int y_normalized) {
     this->angle_led_map[theta].emplace_back(x_normalized, y_normalized);
 }
 
-std::vector<std::tuple<int, int>> CylinderVox::get_slice(int theta) {
-    return angle_led_map[theta];
+std::vector<std::tuple<int, int>> CylinderVox::get_slice(double theta) {
+    return angle_led_map[(int)theta];
 }
